Add data augment methods
